# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TailorInch API",
        "description": "A brief description of your API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/tailors/parameters": {
            "get": {
                "tags": [
                    "Parameters"
                ],
                "summary": "Get all parameters for a specific tailor",
                "operationId": "dd7d74e965a471cea82fbeb9bc51631a",
                "responses": {
                    "200": {
                        "description": "List of parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tailor_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "parameters": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Parameter 1"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "string"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "example": "Some value"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-03T12:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-03T12:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No parameters added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tailor_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "parameters": {
                                            "type": "string",
                                            "example": "No parameters added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token to authorize requests",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}